PROTON-2661 Refactoring buffers and transports for netty 5 support
Changes the ProtonBuffer API to allow for compatability with both netty4
and netty5 and adds a transport in the proton client that will work with
the netty5 snapshot currently available.  Other refactorings and updates
internally to support the netty5 ownershup model for buffers and the
various differences between the two.

+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/DeliveryTagEncoder.java
@@ -46,7 +46,7 @@ public final class DeliveryTagEncoder implements TypeEncoder<DeliveryTag> {
             buffer.writeInt(tagLength);
         } else {
             buffer.writeByte(EncodingCodes.VBIN8);
-            buffer.writeByte(tagLength);
+            buffer.writeByte((byte) tagLength);
         }
 
         value.writeTo(buffer);
